[Fixes] magnitude is a double now

+++ 0x09-csharp-linear_algebra/3-magnitude/3-magnitude.cs
@@ -7,7 +7,7 @@ class VectorMath
     public static double Magnitude(double[] vector)
     {
         int length = vector.Length;
-        int magnitude;
+        double magnitude;
 
         if (length == 2)
         {
